Trying to solve bug with StatActivity. DO NOT MERGE IN MASTER BRANCH!
