+++ app/src/main/java/io/polytech/sportable/activities/settings/ChangeProfile.java
@@ -51,8 +51,6 @@ public class ChangeProfile extends AppCompatActivity {
         }
     }
 
-    boolean flag = true;
-
     @SuppressLint("SetTextI18n")
     public void saveAndQuit(View view) {
