+++ app/src/main/java/io/polytech/sportable/activities/statistics/StatActivity.java
@@ -15,11 +15,18 @@ import io.polytech.sportable.R;
 
 public class StatActivity extends AppCompatActivity {
 
+    StatViewModel local;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        StatViewModel local = new StatViewModel(getApplication());
         setContentView(R.layout.activity_stat);
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        local = new StatViewModel(getApplication(), this);
         ViewPager2 viewPager = findViewById(R.id.stat_view);
         TabLayout tabLayout = findViewById(R.id.tabs);
         viewPager.setAdapter(new FragmentStateAdapter(getSupportFragmentManager(), getLifecycle()) {
@@ -34,37 +41,37 @@ public class StatActivity extends AppCompatActivity {
             @NonNull
             @Override
             public Fragment createFragment(int position) {
-                    if (local.getAllPractices().getValue() != null) {
-                        switch (position) {
-                            case 0:
-                                return new TotalPageFragment(local);
-                            case 1:
-                                return CertainActivityPageFragment.newInstance();
-                            case 2:
-                                return CertainDayFragment.newInstance();
-                            default:
-                                return null;
-                        }
-                    } else {
-                        return NoInformationFragment.newInstance();
+                //if (local.getAllPractices().getValue() != null) {
+                    switch (position) {
+                        case 0:
+                            return new TotalPageFragment(local);
+                        case 1:
+                            return CertainActivityPageFragment.newInstance();
+                        case 2:
+                            return CertainDayFragment.newInstance();
+                        default:
+                            return null;
                     }
-            }
+                } /*else {
+                    return NoInformationFragment.newInstance();
+                }*/
+            //}
         });
         TabLayoutMediator mediator = new TabLayoutMediator(
                 tabLayout, viewPager, true, (tab, position) -> {
-                    switch (position) {
-                        case 0:
-                            tab.setText("Общее");
-                            break;
-                        case 1:
-                            tab.setText("По тренировкам");
-                            break;
-                        case 2:
-                            tab.setText("По дням");
-                            break;
-                    }
-                    viewPager.setCurrentItem(tab.getPosition(), true);
-                });
+            switch (position) {
+                case 0:
+                    tab.setText("Общее");
+                    break;
+                case 1:
+                    tab.setText("По тренировкам");
+                    break;
+                case 2:
+                    tab.setText("По дням");
+                    break;
+            }
+            viewPager.setCurrentItem(tab.getPosition(), true);
+        });
         mediator.attach();
     }
 }
